Add checks for API error responses for browser extension
